Move EntitiesTable
