+++ context/app/static/js/shared-styles/tables/EntitiesTable/EntitiesTable.jsx
