+++ context/app/static/js/shared-styles/tables/EntitiesTable/index.js
