+++ context/app/static/js/shared-styles/tables/EntitiesTable/style.js
