+++ context/app/static/js/shared-styles/tables/EntitiesTable/utils.js
